Added Clue._clean_annotations and added call to Game._parse_clues
The annotations in the Clue methos are now parsed and stored as
a list of tuples of the format (speaker,speech) in Clue.responses
but internal html (such as <i>...</i>) is retained as it can
be considered meaningful data.

Game.__init__ now automatically calls the _parse_clues method so
that Game.clues is populated without the user needing to call an
internal method.